Remove specialized handler and archive struct and restructure handlers pkg.
